Fixing code structure, doc strings, and start testing
